pyproject.toml
[project]
name = "agentic-onboarding-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asana>=5.2.1",
    "atlassian-python-api>=4.0.4",
    "dotenv>=0.9.9",
    "google-api-python-client>=2.178.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.6.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "tabulate>=0.9.0",
]
